requirements.txt
python-dotenv
Flask
Flask-Cors
langchain-google-genai
mysql-connector-python
langchain
SpeechRecognition
gtts
playsound 
pydub